pyproject.toml
[tool.poetry]
name = "flusk"
version = "0.1.0"
description = "DT project"
authors = ["Erik Belak <erik.belak@student.ukf.sk>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3"
flake8 = "^6.0.0"
python-dotenv = "^0"
flask-pymongo = "^2.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"
